studybot/requirements.txt
# StudyBot Agent - Production Dependencies

# Core LangChain (REQUIRED)
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.1.0

# Structured outputs (REQUIRED)
pydantic>=2.0.0

# WebSocket support (optional - only for websocket mode)
websockets>=12.0

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0


# Core
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Dash
dash==2.14.2
dash-bootstrap-components==1.5.0

# Async
asyncio
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0